refactor: moved internal/compiled to frontend/compiled
