+++ frontend/compiled/hint.go
@@ -2,48 +2,12 @@ package compiled
 
 import (
 	"fmt"
-	"io"
 	"math/big"
 	"reflect"
 
-	"github.com/consensys/gnark-crypto/ecc"
-	"github.com/consensys/gnark/backend"
 	"github.com/consensys/gnark/backend/hint"
-	"github.com/consensys/gnark/frontend/schema"
 	"github.com/fxamacker/cbor/v2"
 )
-
-// CS contains common element between R1CS and CS
-type CS struct {
-
-	// number of wires
-	NbInternalVariables int
-	NbPublicVariables   int
-	NbSecretVariables   int
-
-	// logs (added with cs.Println, resolved when solver sets a value to a wire)
-	Logs []LogEntry
-	// debug info contains stack trace (including line number) of a call to a cs.API that
-	// results in an unsolved constraint
-	DebugInfo []LogEntry
-
-	// maps constraint id to debugInfo id
-	// several constraints may point to the same debug info
-	MDebug map[int]int
-
-	Counters []Counter // TODO @gbotrel no point in serializing these
-	// maps wire id to hint
-
-	// a wire may point to at most one hint
-	MHints map[int]*Hint
-
-	Schema *schema.Schema
-
-	// each level contains independent constraints and can be parallelized
-	// it is guaranteed that all dependncies for constraints in a level l are solved
-	// in previous levels
-	Levels [][]int
-}
 
 // Hint represents a solver hint
 // it enables the solver to compute a Wire with a function provided at solving time
@@ -172,47 +136,4 @@ func (h *Hint) UnmarshalCBOR(b []byte) error {
 	h.Inputs = inputs
 	h.Wires = v.Wires
 	return nil
-}
-
-// GetNbVariables return number of internal, secret and public variables
-func (cs *CS) GetNbVariables() (internal, secret, public int) {
-	return cs.NbInternalVariables, cs.NbSecretVariables, cs.NbPublicVariables
-}
-
-// FrSize panics
-func (cs *CS) FrSize() int { panic("not implemented") }
-
-// GetNbCoefficients panics
-func (cs *CS) GetNbCoefficients() int { panic("not implemented") }
-
-// CurveID returns ecc.UNKNOWN
-func (cs *CS) CurveID() ecc.ID { return ecc.UNKNOWN }
-
-// WriteTo panics
-func (cs *CS) WriteTo(w io.Writer) (n int64, err error) { panic("not implemented") }
-
-// ReadFrom panics
-func (cs *CS) ReadFrom(r io.Reader) (n int64, err error) { panic("not implemented") }
-
-// ToHTML panics
-func (cs *CS) ToHTML(w io.Writer) error { panic("not implemtened") }
-
-// GetCounters return the collected constraint counters, if any
-func (cs *CS) GetCounters() []Counter { return cs.Counters }
-
-func (cs *CS) GetSchema() *schema.Schema { return cs.Schema }
-
-func (cs *CS) GetConstraints() [][]string { panic("not implemented") }
-
-// Counter contains measurements of useful statistics between two Tag
-type Counter struct {
-	From, To      string
-	NbVariables   int
-	NbConstraints int
-	CurveID       ecc.ID
-	BackendID     backend.ID
-}
-
-func (c Counter) String() string {
-	return fmt.Sprintf("%s[%s] %s - %s: %d variables, %d constraints", c.BackendID, c.CurveID, c.From, c.To, c.NbVariables, c.NbConstraints)
 }
