+++ frontend/compiled/r1cs_sparse.go
@@ -19,19 +19,40 @@ import (
 	"strings"
 )
 
-// R1C used to compute the wires
-type R1C struct {
-	L, R, O Variable
+// R1CS decsribes a set of SparseR1C constraint
+type SparseR1CS struct {
+	ConstraintSystem
+	Constraints []SparseR1C
 }
 
-func (r1c *R1C) String(coeffs []big.Int) string {
+// GetNbConstraints returns the number of constraints
+func (cs *SparseR1CS) GetNbConstraints() int {
+	return len(cs.Constraints)
+}
+
+// SparseR1C used to compute the wires
+// L+R+M[0]M[1]+O+k=0
+// if a Term is zero, it means the field doesn't exist (ex M=[0,0] means there is no multiplicative term)
+type SparseR1C struct {
+	L, R, O Term
+	M       [2]Term
+	K       int // stores only the ID of the constant term that is used
+}
+
+func (r1c *SparseR1C) String(coeffs []big.Int) string {
 	var sbb strings.Builder
 	sbb.WriteString("L[")
 	r1c.L.string(&sbb, coeffs)
 	sbb.WriteString("] * R[")
 	r1c.R.string(&sbb, coeffs)
-	sbb.WriteString("] = O[")
+	sbb.WriteString("] + M0[")
+	r1c.M[0].string(&sbb, coeffs)
+	sbb.WriteString("] + M1[")
+	r1c.M[1].string(&sbb, coeffs)
+	sbb.WriteString("] + O[")
 	r1c.O.string(&sbb, coeffs)
+	sbb.WriteString("] + K[")
+	sbb.WriteString(coeffs[r1c.K].String())
 	sbb.WriteString("]")
 
 	return sbb.String()
