+++ frontend/cs/plonk/conversion.go
@@ -19,6 +19,7 @@ package plonk
 import (
 	"github.com/consensys/gnark-crypto/ecc"
 	"github.com/consensys/gnark/frontend"
+	"github.com/consensys/gnark/frontend/compiled"
 	"github.com/consensys/gnark/frontend/schema"
 	bls12377r1cs "github.com/consensys/gnark/internal/backend/bls12-377/cs"
 	bls12381r1cs "github.com/consensys/gnark/internal/backend/bls12-381/cs"
@@ -26,14 +27,13 @@ import (
 	bn254r1cs "github.com/consensys/gnark/internal/backend/bn254/cs"
 	bw6633r1cs "github.com/consensys/gnark/internal/backend/bw6-633/cs"
 	bw6761r1cs "github.com/consensys/gnark/internal/backend/bw6-761/cs"
-	"github.com/consensys/gnark/internal/backend/compiled"
 )
 
 func (cs *sparseR1CS) Compile() (frontend.CompiledConstraintSystem, error) {
 
 	res := compiled.SparseR1CS{
-		CS:          cs.CS,
-		Constraints: cs.Constraints,
+		ConstraintSystem: cs.ConstraintSystem,
+		Constraints:      cs.Constraints,
 	}
 	res.NbPublicVariables = len(cs.Public)
 	res.NbSecretVariables = len(cs.Secret)
@@ -121,17 +121,17 @@ HINTLOOP:
 
 	switch cs.CurveID {
 	case ecc.BLS12_377:
-		return bls12377r1cs.NewSparseR1CS(res, cs.Coeffs), nil
+		return bls12377r1cs.NewSparseR1CS(res, cs.builder.Coeffs), nil
 	case ecc.BLS12_381:
-		return bls12381r1cs.NewSparseR1CS(res, cs.Coeffs), nil
+		return bls12381r1cs.NewSparseR1CS(res, cs.builder.Coeffs), nil
 	case ecc.BN254:
-		return bn254r1cs.NewSparseR1CS(res, cs.Coeffs), nil
+		return bn254r1cs.NewSparseR1CS(res, cs.builder.Coeffs), nil
 	case ecc.BW6_761:
-		return bw6761r1cs.NewSparseR1CS(res, cs.Coeffs), nil
+		return bw6761r1cs.NewSparseR1CS(res, cs.builder.Coeffs), nil
 	case ecc.BLS24_315:
-		return bls24315r1cs.NewSparseR1CS(res, cs.Coeffs), nil
+		return bls24315r1cs.NewSparseR1CS(res, cs.builder.Coeffs), nil
 	case ecc.BW6_633:
-		return bw6633r1cs.NewSparseR1CS(res, cs.Coeffs), nil
+		return bw6633r1cs.NewSparseR1CS(res, cs.builder.Coeffs), nil
 	default:
 		panic("unknown curveID")
 	}
