--- internal/backend/compiled/r1cs_sparse.go
@@ -1,26 +1,0 @@
-// Copyright 2020 ConsenSys AG
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package compiled
-
-// R1CS decsribes a set of SparseR1C constraint
-type SparseR1CS struct {
-	CS
-	Constraints []SparseR1C
-}
-
-// GetNbConstraints returns the number of constraints
-func (cs *SparseR1CS) GetNbConstraints() int {
-	return len(cs.Constraints)
-}
