+++ internal/generator/backend/template/representations/r1cs.go.tmpl
@@ -9,7 +9,7 @@ import (
 	"github.com/fxamacker/cbor/v2"
 
 	"github.com/consensys/gnark/internal/backend/ioutils"
-	"github.com/consensys/gnark/internal/backend/compiled"
+	"github.com/consensys/gnark/frontend/compiled"
 	"github.com/consensys/gnark/frontend/schema"
 	"github.com/consensys/gnark/backend"
 	"github.com/consensys/gnark/backend/witness"
