+++ internal/generator/backend/template/representations/r1cs.sparse.go.tmpl
@@ -11,7 +11,7 @@ import (
 	"math"
 	
 	"github.com/consensys/gnark/internal/backend/ioutils"
-	"github.com/consensys/gnark/internal/backend/compiled"
+	"github.com/consensys/gnark/frontend/compiled"
 	"github.com/consensys/gnark/backend"
 	"github.com/consensys/gnark/frontend/schema"
 	"github.com/consensys/gnark/backend/witness"
