+++ internal/generator/backend/template/representations/solution.go.tmpl
@@ -6,7 +6,7 @@ import (
 	"sync/atomic"
 
     "github.com/consensys/gnark/backend/hint"
-    "github.com/consensys/gnark/internal/backend/compiled"
+    "github.com/consensys/gnark/frontend/compiled"
 	"github.com/consensys/gnark/internal/utils"
 	"github.com/consensys/gnark/frontend/schema"
