+++ internal/generator/backend/template/zkpschemes/groth16/groth16.setup.go.tmpl
@@ -4,7 +4,7 @@ import (
 	{{ template "import_backend_cs" . }}
 	{{ template "import_fft" . }}
 	"github.com/consensys/gnark-crypto/ecc"
-	"github.com/consensys/gnark/internal/backend/compiled"
+	"github.com/consensys/gnark/frontend/compiled"
 	"math/big"
 	"math/bits"
 )
